--- Assets/Animations/CinemachineAnimation.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cd36c815ba925bd43b121cdf951fc972
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
