+++ Assets/Evidences/Evidence 1.asset
@@ -20,5 +20,5 @@ MonoBehaviour:
   Artwork: {fileID: 2100000, guid: fcac397647bac7e4fb88c99de4162c58, type: 2}
   ArtworkXsize: 0
   ArtworkYsize: 0
-  Object: {fileID: 3242774187239063980, guid: 226502d392ed5b048971e13fa89155a4, type: 3}
+  Object: {fileID: 0}
   evidenceType: 1
