--- Assets/Scripts/DragLine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f787fe8abf2e65c4f997022f2f2466e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
