+++ Assets/Scripts/PlayerController.cs
@@ -80,6 +80,7 @@ public class PlayerController : MonoBehaviour
 
 
     }
+
     private void Inspect()
     {
